Fix large-image crashes and clean compiler warnings (#221)
A large image could evict a bitmap that the document still needed to draw, crashing Tinta with an access violation. I gave the layout and cache independent bitmap ownership, corrected pixel-memory accounting, and shared bounded WIC decoding between local and downloaded images. Invalid or excessive raster inputs use the existing alt-text placeholder; HTML and DOCX retain original image bytes. Reported by @glance02 in #220.

I also removed unused variables and arguments, renamed shadowed locals, and made narrowing conversions explicit without suppressing compiler warnings.

Validation: all-target Release build with zero warnings; 21/21 CTests; reproduced 3.7.0 crash cases now pass; localhost HTTP decoding, lightbox, normal/narrow Paper/Midnight views, mixed headings/tables/quotes/math/code, and HTML/DOCX/PDF exports checked. Tests generate large image assets outside Git. Details are in tests/image-cache-validation.md and tests/compiler-warning-cleanup-validation.md.

Addresses #220. The release will close the issue after v3.7.1 is published.

+++ tests/file_path_tests.cpp
@@ -41,7 +41,8 @@ void clickTabCopy(App& app) {
     // Copy file path follows the four close actions and their separator.
     int x = static_cast<int>(app.tabMenuX + 20);
     int y = static_cast<int>(app.tabMenuY + dpi(app, 6+4*30+9+15));
-    check(tabMenuItemAt(app, x, y) == 4, "tab Copy file path remains next to Reveal");
+    check(tabMenuItemAt(app, static_cast<float>(x), static_cast<float>(y)) == 4,
+          "tab Copy file path remains next to Reveal");
     check(tabMenuMouseDown(app, app.hwnd, x, y), "tab path action consumes its click");
     check(!app.showTabMenu, "tab Copy file path closes the menu");
 }
@@ -100,7 +101,7 @@ void titleMenuCases(bool testClipboard) {
                         }
                         app.tabs.resize(1);  // last companion closed
                         renderStrip(app);
-                        int x = (int)dpi(app, edit ? 60 : 48);
+                        int x = (int)dpi(app, edit ? 60.0f : 48.0f);
                         int y = (int)dpi(app, 20);
                         check(!tabStripVisible(app) && tabContextMenuIndexAt(app, (float)x, (float)y) == 0,
                               "the lone title has a tab context target after closing a companion");
